fix: Move write buffer data types to write_buffer crate
